Network.UDP

Description

Best current practice library for UDP clients and servers.

Efficient receiving function without memory copy

Proper buffer size

Type-safe APIs

TCP-like APIs (creating a UDP connection from a listing socket) in the server side

Auto migration (network interface selection) in the client side

The recv family in Network.Socket.ByteString uses createAndTrim internaly. So, one buffer is allocated before corresponding system calls are called. Then another buffer is allocated according to the input size and the input is copied. Receiving functions provided by this library uses createUptoN to avoid the memory copy.

Recent application protocols are designed to avoid IP fragmentation. So, the UDP payload size is never over 1,500. This library uses 2,048 for the buffer size. This size ensures no global locking when allocating ByteString (i.e. a buffer).

To know the background of TCP-like API in the server side, see:

https://kazu-yamamoto.hatenablog.jp/entry/2022/02/25/153122

To know the background of auto migration in the client side, see:

https://kazu-yamamoto.hatenablog.jp/entry/2021/06/29/134930

https://www.iij.ad.jp/en/dev/iir/pdf/iir_vol52_focus2_EN.pdf (Sec 3.9)

Sockets used by clients and servers after accept

data UDPSocket

A UDP socket which are used with recv and send.

Constructors

UDPSocket

Fields

udpSocket :: Socket

peerSockAddr :: SockAddr

Used for a unconnected socket naturally. Used for a connected sockdet for checking

connected :: Bool

clientSocket :: HostName -> ServiceName -> Bool -> IO UDPSocket

Creating a unconnected UDP socket.

recv :: UDPSocket -> IO ByteString

Receiving data with a UDP socket. If the socket is connected, recv() is called. Otherwise, recvfrom() is called.

recvBuf :: UDPSocket -> Ptr Word8 -> Int -> IO Int

Receiving data in a buffer with a UDP socket. If the socket is connected, recv() is called. Otherwise, recvfrom() is called.

send :: UDPSocket -> ByteString -> IO ()

Sending data with a UDP socket. If the socket is connected, send() is called. Otherwise, sento() is called.

sendBuf :: UDPSocket -> Ptr Word8 -> Int -> IO ()

Sending data in a buffer with a UDP socket. If the socket is connected, send() is called. Otherwise, sento() is called.

Server's wildcard socket

data ListenSocket

A listening socket for UDP which can be used for recvFrom and sendTo. Optionally, a connected UDP socket can be created with accept as an emulation of TCP.

Constructors

ListenSocket

Fields

listenSocket :: Socket

mySockAddr :: SockAddr

wildcard :: Bool

True for wildcard. False for interface-specific.

serverSocket :: (IP, PortNumber) -> IO ListenSocket

Creating a listening UDP socket.

data ClientSockAddr

A client socket address from the server point of view.

Constructors

ClientSockAddr SockAddr [Cmsg]

recvFrom :: ListenSocket -> IO (ByteString, ClientSockAddr)

Receiving data with a listening UDP socket. For a wildcard socket, recvmsg() is called. For an interface specific socket, recvfrom() is called.

sendTo :: ListenSocket -> ByteString -> ClientSockAddr -> IO ()

Sending data with a listening UDP socket. For a wildcard socket, sendmsg() is called. For an interface specific socket, sento() is called.

Server's connected socket

accept :: ListenSocket -> ClientSockAddr -> IO UDPSocket

Creating a connected UDP socket like TCP's accept().

Closing

stop :: ListenSocket -> IO ()

Closing a socket.

close :: UDPSocket -> IO ()

Closing a socket.

Misc

natRebinding :: UDPSocket -> IO UDPSocket

Emulation of NAT rebiding in the client side. This is mainly used for test purposes.